refactor: match swagger api and controllers
