+++ src/models/IExpense.ts
@@ -16,18 +16,3 @@ export default interface IExpense {
         icon_url: string
     }
 }
-
-export interface IExpensePeriodYearMonth {
-    year_month: string,
-} 
-export interface IExpensePeriodRangeDates {
-    start_date: string,
-    end_date: string
-}
-export interface IExpensePeriods {
-    period: {
-        year_month?: string,
-        start_date?: string,
-        end_date?: string
-    }
-}
