+++ src/store/Controllers/ExpensesController/ExpensesControllerInterfaces.ts
@@ -1,5 +1,6 @@
-import IExpense, { IExpensePeriodRangeDates, IExpensePeriodYearMonth } from "@models/IExpense"
-import IUser, { IExtendedUser } from "@models/IUser"
+import IExpense from "@models/IExpense"
+import { IPeriodRangeDates, IPeriodYearMonth } from "@models/IPeriod";
+import IUser from "@models/IUser"
 
 
 // create & update 
@@ -8,14 +9,7 @@ export interface IExpenseByGroupResponse extends IExpense {
 }
 
 export interface IGetExpensesBody {
-    period: IExpensePeriodYearMonth | IExpensePeriodRangeDates
-}
-export interface IGetCurrentUserDailyExpensesResponse {
-    date: string,
-    amount: number
-}
-export interface IGetCurrentUserDailyExpensesBody {
-    period: IExpensePeriodYearMonth | IExpensePeriodRangeDates
+    period: IPeriodYearMonth | IPeriodRangeDates
 }
 
 export interface ICreateExpenseByGroupBody {
@@ -32,41 +26,6 @@ export interface IUpdateExpenseByGroupBody extends ICreateExpenseByGroupBody {
 
 export interface IGetExpensesByGroupBody {
     group_id: number,
-    period: IExpensePeriodYearMonth | IExpensePeriodRangeDates
+    period: IPeriodYearMonth | IPeriodRangeDates
 }
 
-export interface IGetTotalExpensesBody {
-    group_id: number,
-    period: IExpensePeriodYearMonth | IExpensePeriodRangeDates
-}
-export interface IGetTotalExpensesResponse {
-    amount: number,
-    percentage_increase: number
-}
-
-export interface IGetCurrentGroupSpendersResponse extends IExtendedUser { }
-export interface IGetCurrentGroupSpendersBody {
-    group_id: number,
-    period: IExpensePeriodYearMonth | IExpensePeriodRangeDates
-}
-
-export interface IGetTotalExpensesBody {
-    group_id: number, 
-    period: IExpensePeriodYearMonth | IExpensePeriodRangeDates
-}
-export interface IGetTotalExpensesResponse {
-    amount: number,
-    percentage_increase: number
-}
-
-export interface IGetCurrentGroupSpendersResponse {
-    id: number,
-    first_name: string,
-    last_name: string,
-    picture: string,
-    amount: number
-}
-export interface IGetCurrentGroupSpendersBody {
-    group_id: number, 
-    period: IExpensePeriodYearMonth | IExpensePeriodRangeDates
-}
