+++ src/store/Controllers/UserController/UserControllerInterfaces.ts
@@ -7,34 +7,13 @@ export interface IGetCurrentUserInfo extends IUser{}
 export interface IGetCurrentUserBalance {
     balance: number
 }
-
-export interface IGetUsersFromGroupResponse {
-    items: [
-        {
-            users_group: [
-                {
-                    user: IUser
-                    status: string,
-                    date_join: string
-                }
-            ]
-        }
-    ],
-    total: 0,
-    page: 0,
-    size: 0,
-    pages: 0
+export interface IGetCurrentUserDailyExpensesResponse {
+    date: string,
+    amount: number
 }
-export interface IGetTotalExpensesResponse {
-    amount: number,
-    percentage_increase: number
-}
-export interface IGetTotalReplenishmentsResponse extends IGetTotalExpensesResponse{}
-export interface IGetTotalExpensesBody {
+export interface IGetCurrentUserDailyExpensesBody {
     period: IPeriodYearMonth | IPeriodRangeDates
 }
-export interface IGetTotalReplenishmentsBody extends IGetTotalExpensesBody { }
-
 export interface IGetUserExpensesByGroupResponse {
     group_id: number,
     group_title: string,
@@ -45,6 +24,16 @@ export interface IGetUserExpensesByGroupBody {
     group_id: number,
     period: IPeriodYearMonth | IPeriodRangeDates;
 }
+
+export interface IGetTotalExpensesResponse {
+    amount: number,
+    percentage_increase: number
+}
+export interface IGetTotalReplenishmentsResponse extends IGetTotalExpensesResponse{}
+export interface IGetTotalExpensesBody {
+    period: IPeriodYearMonth | IPeriodRangeDates
+}
+export interface IGetTotalReplenishmentsBody extends IGetTotalExpensesBody { }
 
 export interface IGetUserExpensesByCategoryResponse extends ICategoryAmount {}
